chore: rename temporary_messages module to temporary_session
